Make one frame price a BigDecimal instead of Double

+++ forex-mtl/src/main/scala/forex/services/rates/Protocol.scala
@@ -19,7 +19,7 @@ object Protocol {
   final case class OneFrameRate(
     from: String,
     to: String,
-    price: Double,
+    price: BigDecimal,
     time_stamp: OffsetDateTime
   )
